requirements.txt
streamlit
sqlalchemy
pymysql
python-dotenv
euriai
psycopg2-binary
plotly